+++ crates/holochain_types/src/app/app_manifest/app_manifest_v1.rs
@@ -38,12 +38,15 @@ pub struct AppManifestV1 {
     /// The roles that need to be filled (by DNAs) for this app.
     pub roles: Vec<AppRoleManifest>,
 
-    /// Declares that the app should be installed without the need to
-    /// specify membrane proofs at installation time. The membrane proofs
-    /// will instead be provided later by [`AppRequest::ProvideMemproofs`]
+    /// Declares that the app may be installed without the need to
+    /// specify membrane proofs at installation time. If memproofs are not
+    /// provided at install time, they must be provided later by the
+    /// [`AppRequest::ProvideMemproofs`] call. If memproofs are provided
+    /// at install time, the app will be installed as normal, without the
+    /// special deferred memproof flow.
     #[serde(default)]
     #[builder(default)]
-    pub membrane_proofs_deferred: bool,
+    pub allow_deferred_memproofs: bool,
 }
 
 /// Description of an app "role" defined by this app.
@@ -197,7 +200,7 @@ impl AppManifestV1 {
             name,
             roles,
             description: _,
-            membrane_proofs_deferred: _,
+            allow_deferred_memproofs: _,
         } = self;
         let roles = roles
             .into_iter()
@@ -296,7 +299,7 @@ pub mod tests {
             name: "Test app".to_string(),
             description: Some("Serialization roundtrip test".to_string()),
             roles,
-            membrane_proofs_deferred: false,
+            allow_deferred_memproofs: false,
         }
     }